requirements.txt
# Slack Bot
slack-bolt>=1.18.0
slack-sdk>=3.21.0

# Browser Automation
playwright>=1.40.0

# PDF Processing and OCR
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Date parsing
dateparser>=1.2.0

# Text matching
rapidfuzz>=3.5.0

# Environment management
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0
aiofiles>=23.2.0

# Scheduling
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
